backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
asyncpg
qdrant-client
pydantic-settings
alembic
redis
arq
httpx
openai
langchain
langchain-community
